requirements.txt
fastapi==0.115.6
pydantic==2.9.2
streamlit==1.36.0
sqlite3
uvicorn
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
ollama==0.1.7
langgraph==0.4.7
langchain-core==0.3.17
plotly==5.22.0
pandas==2.2.2
asyncio